backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg[binary]>=3.2,<4.0
cryptography>=43.0,<44.0
reportlab>=4.2,<5.0
pytest>=8.2,<9.0
pytest-django>=4.8,<5.0
gunicorn>=23.0,<24.0
